# Optimize the tail base64 encode for Icelake