For inputs of smaller lengths like <64 bytes, the icelake implementation of base64 encode calls the scalar function to process. This functionality is implemented in this PR which saves some clock cycles for small input encoding.

Before the change, running the base64 benchmark on 48 byte input:
```
# current system detected as icelake.
# loading files: .
# volume: 48 bytes
# max length: 48 bytes
# number of inputs: 1
# encode
memcpy                                   :   2.81 GB/s  16.34 GHz   5.82 c/b   5.21 i/b   0.90 i/c
libbase64                                :   1.10 GB/s   9.65 GHz   8.80 c/b  24.52 i/b   2.79 i/c
simdutf::icelake                         :   1.45 GB/s  11.07 GHz   7.62 c/b  16.67 i/b   2.19 i/c
simdutf::haswell                         :   1.87 GB/s  12.67 GHz   6.79 c/b  12.33 i/b   1.82 i/c
simdutf::westmere                        :   2.08 GB/s  13.55 GHz   6.52 c/b  10.44 i/b   1.60 i/c
simdutf::fallback                        :   1.50 GB/s  11.48 GHz   7.64 c/b  16.27 i/b   2.13 i/c
```

With this change:
```
# current system detected as icelake.
# loading files: .
# volume: 48 bytes
# max length: 48 bytes
# number of inputs: 1
# encode
memcpy                                   :   2.81 GB/s  16.53 GHz   5.89 c/b   5.21 i/b   0.88 i/c
libbase64                                :   1.10 GB/s   9.63 GHz   8.79 c/b  24.52 i/b   2.79 i/c
simdutf::icelake                         :   2.22 GB/s  14.09 GHz   6.36 c/b   6.23 i/b   0.98 i/c
simdutf::haswell                         :   1.89 GB/s  12.80 GHz   6.79 c/b  12.33 i/b   1.82 i/c
simdutf::westmere                        :   2.08 GB/s  13.57 GHz   6.54 c/b  10.44 i/b   1.60 i/c
simdutf::fallback                        :   1.49 GB/s  11.21 GHz   7.51 c/b  16.27 i/b   2.17 i/c
```