@lemire 2024-11-03
I manually verified that for tiny inputs, this PR improves the performance substantially for the AVX-512 kernel. Merging.